refactor: devide queries for calc
